requirements.txt
Flask==3.0.0
google-cloud-storage==2.13.0
gunicorn==21.2.0
google-cloud-bigquery==3.13.0
APScheduler==3.7.0
